Moving git tools to its own section.
