+++ kernel/git/apis/cli/gitpython.py
@@ -22,11 +22,11 @@ from git import Remote as RemoteAPI
 from git import Repo as RepoAPI
 from overrides import overrides
 
+from kernel.git.apis.cli import Git as IGit
+from kernel.git.apis.cli import GitError
+from kernel.git.apis.cli import Remote as IRemote
+from kernel.git.apis.cli import Repository as IRepository
 from kernel.tools.fs import Dir, File
-from kernel.tools.git import Git as IGit
-from kernel.tools.git import GitError
-from kernel.tools.git import Remote as IRemote
-from kernel.tools.git import Repository as IRepository
 from kernel.tools.urls import URL
 
 LOG = logging.getLogger(__name__)
